Move delta-related code to separate module
