+++ src/main.rs
@@ -1,5 +1,5 @@
 use clap::Parser;
-use lakehouse_loader::{do_main, Cli, DataLoadingError};
+use lakehouse_loader::{do_main, error::DataLoadingError, Cli};
 
 #[tokio::main]
 async fn main() -> Result<(), DataLoadingError> {
